refactored layout of divide_string package and added tests
