+++ task/divide_string/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+
+	"problems/task/divide_string/pkg/divide_string"
+)
+
+func main() {
+	fmt.Println(divide_string.DivStr("1234"))
+}
