+++ task/divide_string/pkg/divide_string/divide_string_test.go
@@ -1,0 +1,39 @@
+package divide_string
+
+import (
+	"fmt"
+	"reflect"
+	"testing"
+)
+
+func TestDivStr(t *testing.T) {
+	testCases := []struct {
+		input    string
+		expected []string
+	}{
+		{
+			input:    "12345",
+			expected: []string{"12", "34", "5_"},
+		},
+
+		{
+			input:    "12345678",
+			expected: []string{"12", "34", "56", "78"},
+		},
+
+		{
+			input:    "1",
+			expected: []string{"1_"},
+		},
+	}
+
+	for i, testCase := range testCases {
+		t.Run("case_"+fmt.Sprint(i), func(t *testing.T) {
+			ans := DivStr(testCase.input)
+
+			if !reflect.DeepEqual(ans, testCase.expected) {
+				t.Errorf("expected = %s, actual = %s", testCase.expected, ans)
+			}
+		})
+	}
+}
