+++ task/math/is_proper_number.go
