+++ task/permutations/permutations.go
