+++ task/slices/del_from_slice.go
