+++ task/writing_cases/toWeirdCase.go
